Adding modified generic list view that would filter based on given querystring from request.
