+++ myportfolio/core/views.py
@@ -7,42 +7,7 @@ from django.conf import settings
 from blog.models import Post
 from projects.models import Project
 
-from .utils import group_pagination
-
 # Create your views here.
-
-class ModifiedPaginateListView(generic.ListView):
-    # TODO: Check if this is needed.
-    def get(self, request, *args, **kwargs):
-        self.object_list = self.get_queryset()
-        allow_empty = self.get_allow_empty()
-        
-        if not allow_empty:
-            # When pagination is enabled and object_list is a queryset,
-            # it's better to do a cheap query than to load the unpaginated
-            # queryset in memory.
-            if self.get_paginate_by(self.object_list) is not None and hasattr(self.object_list, 'exists'):
-                is_empty = not self.object_list.exists()
-            else:
-                is_empty = not self.object_list
-                if is_empty:
-                    raise Http404(_("Empty list and '%(class_name)s.allow_empty' is False.") % {
-                        'class_name': self.__class__.__name__,
-                        })
-        # TODO: By given page_num, determine the current group_number
-
-        context = self.get_context_data()
-        if context['is_paginated']:
-            page_obj = context['page_obj']
-            num_pages = page_obj.paginator.num_pages
-
-            # give default value of 1
-            group_num = int(request.GET.get('group_num', 1))
-
-            context.update(group_pagination(num_pages, group_num))
-
-        return self.render_to_response(context)    
-
 
 class HomeView(generic.TemplateView):
     template_name = 'home.html'
