+++ templates/myportfolio/_navigation.html
@@ -18,7 +18,7 @@
             <a href="#!" style="padding: 0 16px;">Blog<i class="material-icons">arrow_drop_down</i></a>
             <div>
               <ul>
-                <li><a href="{% url 'post-list' %}">View Posts</a></li>
+                <li><a href="{% url 'post-list' %}?refresh=true">View Posts</a></li>
                 <li><a href="{% url 'post-form' %}">Create Post</a></li>
               </ul>
             </div>
@@ -41,7 +41,7 @@
       </li>   
       <li><a href="{% url 'logout' %}">Logout</li> 
     {% else %}
-    <li><a href="{% url 'post-list' %}">Blog</a></li>
+    <li><a href="{% url 'post-list' %}?refresh=true">Blog</a></li>
     <li><a class="waves-effect" href="{% url 'project-list' %}">Projects</a></li>
     {% endif %}
 </ul>
